Add basic param validation in query bindings
